refactor(tests): reorganized test-entities
